requirements.txt
dify_plugin>=0.4.0,<0.7.0
tablestore-agent-storage==1.0.5
